#20401 #21402 Api for address template - automapper Reviewed by: Marta Dycjan

+++ src/KnightFrank.Antares.Domain/AddressFieldDefinition/AddressFieldDefinitionResult.cs
@@ -10,8 +10,8 @@
 
         public bool Regex { get; set; }
 
-        public int RowOrder { get; set; }
+        public short RowOrder { get; set; }
 
-        public int ColumnOrder { get; set; }
+        public short ColumnOrder { get; set; }
     }
 }
